+++ packages/ag-grid-community/src/interfaces/iModule.ts
@@ -191,6 +191,7 @@ export type EnterpriseModuleName =
     | 'Pivot'
     | 'RangeSelection'
     | 'RichSelect'
+    | 'RichSelectAsyncRequests'
     | 'RowNumbers'
     | 'RowGrouping'
     | 'RowGroupingPanel'
@@ -273,6 +274,7 @@ export type AgModuleName =
     | 'PivotModule'
     | 'RangeSelectionModule'
     | 'RichSelectModule'
+    | 'RichSelectAsyncRequestsModule'
     | 'RowNumbersModule'
     | 'RowGroupingModule'
     | 'RowGroupingPanelModule'
