+++ packages/ag-grid-community/src/validation/enterpriseModuleNames.ts
@@ -23,6 +23,7 @@ export const ENTERPRISE_MODULE_NAMES: Record<EnterpriseModuleName, 1> = {
     Pivot: 1,
     RangeSelection: 1,
     RichSelect: 1,
+    RichSelectAsyncRequests: 1,
     RowNumbers: 1,
     RowGrouping: 1,
     RowGroupingPanel: 1,
