+++ packages/ag-grid-enterprise/src/richSelect/richSelectAsyncRequestsModule.ts
@@ -1,0 +1,11 @@
+import type { _ModuleWithoutApi } from 'ag-grid-community';
+
+import { VERSION } from '../version';
+
+/**
+ * @internal
+ */
+export const RichSelectAsyncRequestsModule: _ModuleWithoutApi = {
+    moduleName: 'RichSelectAsyncRequests',
+    version: VERSION,
+};
